test: criar testes automatizados para classe Pedido.java

+++ src/main/java/br/com/infuse/desafio/infrastructure/pedido/models/ListarPedidoResponse.java
@@ -21,7 +21,7 @@ public class ListarPedidoResponse {
     @JsonProperty("quantidade")
     private Integer quantidade;
     @JsonProperty("codigo_cliente")
-    private Long codigoCliente;
+    private Integer codigoCliente;
     @JsonProperty("nome")
     private String nome;
